+++ components/dashboard/PasswordManager/PasswordTable.module.scss
