+++ components/dashboard/PasswordManager/PasswordTable/FilterInput.module.scss
@@ -1,5 +1,5 @@
-@use "../../../../styles/colors.module";
-@use "../../../../styles/mixins";
+@use "styles/colors.module";
+@use "styles/mixins";
 
 $input_height: 40;
