+++ components/dashboard/PasswordManager/PasswordTable/TableEntry.module.scss
@@ -1,6 +1,6 @@
-@use "../../../../styles/colors.module";
-@use "../../../../styles/animations";
-@use "../../../../styles/mixins";
+@use "styles/colors.module";
+@use "styles/animations";
+@use "styles/mixins";
 
 $avatar_size: 50px;
