+++ components/dashboard/PasswordManager/PasswordTable/TableEntry/ClosedEntry.tsx
