+++ components/dashboard/PasswordManager/PasswordTable/TableEntry/ExpandedEntry.tsx
