+++ components/dashboard/PasswordManager/PasswordTable/TableEntry/ExpandedEntry/DeleteModal.module.scss
