+++ components/dashboard/PasswordManager/PasswordTable/TableEntry/ExpandedEntry/DeleteModal.tsx
