+++ components/dashboard/PasswordManager/PasswordTable/TableEntry/ExpandedEntry/EntryInput.module.scss
