+++ components/dashboard/PasswordManager/SidePanel.module.scss
