+++ components/dashboard/PasswordManager/SidePanel.tsx
